stack-create: Fix bug with stack alignment for ld.so
ld.so expects the stack start to be always 16-byte aligned. In
particular, it uses this instruction (movaps) which only works with
16-byte aligned memory.

For now, we get around this problem by calling the kernel loader with 1
argument. The kernel automatically gives us a 16-byte aligned stack.
Since we inherit the stack from the kernel loader, we also get a 16-byte
aligned stack for ld.so.

In the previous method, the kernel loader's stack was aligned for
16-bytes, but by pushing an extra argument on ld.so's stack, we
ended up destroying the alignment. This would cause the movaps
instruction to raise a general protection exception (#GP), which
translates to a segfault for the user process.

With this fix, the full solution works -- end-to-end! :-)